automation/requirements.txt
isla-solver == 1.14.2
lxml == 4.9.3
selenium == 4.10.0
selenium-wire == 5.1.0
pyyaml == 6.0.1